Split action files + fix responsivity
