+++ src/core/InsightContainer/Action/index.js
@@ -3,41 +3,32 @@ import { PropTypes } from 'prop-types';
 
 import { makeStyles } from '@material-ui/styles';
 
-import { IconButton, Typography, Grid, Popover } from '@material-ui/core';
+import { Typography, Grid, Popover } from '@material-ui/core';
 
-import shareIcon from '../assets/icons/network-connection.svg';
-import embedIcon from '../assets/icons/code.svg';
-import downloadIcon from '../assets/icons/download.svg';
-import compareIcon from '../assets/icons/compare.svg';
-import showIcon from '../assets/icons/tablet-reader.svg';
-import BlockLoader from '../BlockLoader';
+import ActionButton from './ActionButton';
 
-const useStyles = makeStyles(({ breakpoints }) => ({
+import shareIcon from '../../assets/icons/network-connection.svg';
+import embedIcon from '../../assets/icons/code.svg';
+import downloadIcon from '../../assets/icons/download.svg';
+import compareIcon from '../../assets/icons/compare.svg';
+import showIcon from '../../assets/icons/tablet-reader.svg';
+import BlockLoader from '../../BlockLoader';
+
+const useStyles = makeStyles({
   root: {
+    width: '100%',
+    maxWidth: '21.75rem',
     overflow: 'hidden',
     backgroundColor: 'white',
     borderRadius: '0.625rem',
-    marginTop: '0.6rem',
-    width: '100%',
-    maxWidth: '21.75rem',
-    [breakpoints.up('md')]: {
-      width: '16.3125rem' // .75 of lg
-    },
-    [breakpoints.up('lg')]: {
-      width: '21.75rem'
-    }
+    marginTop: '0.6rem'
   },
   actionButton: {
     borderRadius: '0',
     '&:hover': {
       backgroundColor: 'transparent'
     },
-    [breakpoints.up('md')]: {
-      padding: '0.625rem 0.25rem'
-    },
-    [breakpoints.up('lg')]: {
-      padding: '0.625rem'
-    }
+    padding: '0.625rem'
   },
   actionButtonText: {
     fontSize: '0.6rem',
@@ -61,68 +52,7 @@ const useStyles = makeStyles(({ breakpoints }) => ({
   embedButton: {},
   showDataButton: {},
   downloadButton: {}
-}));
-
-function ActionButton({
-  children,
-  onClick,
-  gaEvents: { gaOn, gaEventAction, gaEventCategory, gaEventLabel },
-  classes: propClasses,
-  ...props
-}) {
-  const classes = useStyles({ ...propClasses, props });
-  return (
-    <IconButton
-      className={classes.actionButton}
-      onClick={onClick}
-      ga-on={gaOn}
-      ga-event-category={gaEventCategory}
-      ga-event-action={gaEventAction}
-      ga-event-label={gaEventLabel}
-      {...props}
-    >
-      <Grid
-        component="span"
-        container
-        direction="column"
-        justify="space-between"
-        alignItems="center"
-        className={classes.iconGrid}
-      >
-        {children}
-      </Grid>
-    </IconButton>
-  );
-}
-
-ActionButton.propTypes = {
-  onClick: PropTypes.func,
-  classes: PropTypes.shape({
-    actionButton: PropTypes.string,
-    iconGrid: PropTypes.string
-  }),
-  children: PropTypes.oneOfType([
-    PropTypes.arrayOf(PropTypes.node),
-    PropTypes.node
-  ]).isRequired,
-  gaEvents: PropTypes.shape({
-    gaOn: PropTypes.string,
-    gaEventCategory: PropTypes.string,
-    gaEventAction: PropTypes.string,
-    gaEventLabel: PropTypes.string
-  })
-};
-
-ActionButton.defaultProps = {
-  classes: undefined,
-  onClick: null,
-  gaEvents: {
-    gaOn: undefined,
-    gaEventCategory: undefined,
-    gaEventAction: undefined,
-    gaEventLabel: undefined
-  }
-};
+});
 
 const EmbedCodeTextArea = ({ code }) => {
   const ref = createRef();
